MSB-3896 - Added performance cost to the well WellCoordinatesResponse API

+++ src/main/java/com/moblize/ms/dailyops/dto/Cost.java
@@ -4,7 +4,11 @@ package com.moblize.ms.dailyops.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 import org.springframework.data.mongodb.core.mapping.Document;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -23,11 +27,11 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public class Cost {
 
     @JsonProperty("afe")
-    public Double afe;
+    public Integer afe;
     @JsonProperty("perFt")
-    public Double perFt;
+    public Integer perFt;
     @JsonProperty("perLatFt")
-    public Double perLatFt;
+    public Integer perLatFt;
     @JsonProperty("total")
     public Integer total;
